==> Replication files/2_Multiple_OLS.Rmd <==
---
title: "Multiple OLS Regressions"
author: "Michelle Schultze"
date: "Sept 21, 2024"
output:
  pdf_document:
    toc: yes
  html_document:
    toc: yes
    toc_float: yes
---
---

##Introduction

```{r packages, message=FALSE, warning=FALSE}
library(tidyverse)
library(dplyr)
library(readr)
library(readxl)
library(naniar)
#install.packages("readr", repos = "http://cran.us.r-project.org")
#install.packages("microsynth", repos = "http://cran.us.r-project.org")
#install.packages("LowRankQP", repos = "http://cran.us.r-project.org")
library("readr")
library("microsynth")
library("LowRankQP")
#install.packages("plm")
library(plm)
library(lmtest)
library(modelsummary)
library(kableExtra)

setwd('/Users/michelle/Documents/ukr2024/Replication files')
data <- read_csv("full_data.csv")
```

```{r}
# GDP deflator = nominal/real * 100
# nominal = real * deflator/100

data <- data %>%
  mutate(total_GDP_n_ID = (`GDP, PPP (constant 2017 international $)` * `GDP deflator (base year varies by country)` /100))

#before reporting, check that this is the proper way to do this. (says "inflation" in the title of the variable which doesn't sound like what we want? But this is supposed to be the GDP deflator.. )
# chatgpt confirms √: mutate(Nominal_GDP = GDP_PPP * (GDP_deflator / 100))
```

Interpretation is hard here because the coefficients on percent change in passengers will be too small for interpretation (representing each 1% increase in passengers). Let's divide passengers/100 so that we have them in fractional terms. The coefficient will describe every additional 100% increase in passengers (where -50% is a decrease by half of the previous value, 0% there is no change, 100% is a doubling, and 200% is a tripling)
```{r}
data <- data %>%
  mutate(passengers.in.change.pct = passengers.in.change.pct / 100)
```

##IHS-Transformed Real FDI

```{r}
models <- NA
models <- list()
hsd_testing <- NA
hsd_testing <- list()

model.1a <- lm(ihs_FDI_r ~ passengers.in.change.pct, data = data)
models[["IHS-Transformed Real FDI"]][["1a"]] <- model.1a
hsd_testing[["IHS-Transformed Real FDI"]][["1a"]] <- bptest(model.1a)

model.2a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle), data = data)
models[["IHS-Transformed Real FDI"]][["2a"]] <- model.2a
hsd_testing[["IHS-Transformed Real FDI"]][["2a"]] <- bptest(model.2a)

model.3a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r), data = data)
models[["IHS-Transformed Real FDI"]][["3a"]] <- model.3a
hsd_testing[["IHS-Transformed Real FDI"]][["3a"]] <- bptest(model.3a)

model.4a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r_pc), data = data)
models[["IHS-Transformed Real FDI"]][["4a"]] <- model.4a
hsd_testing[["IHS-Transformed Real FDI"]][["4a"]] <- bptest(model.4a)

model.5a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log1p(total_GDP_n_ID)), data = data)
models[["IHS-Transformed Real FDI"]][["5a"]] <- model.5a
hsd_testing[["IHS-Transformed Real FDI"]][["5a"]] <- bptest(model.5a)

model.6a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log1p(`GDP.current.US$`)), data = data)
models[["IHS-Transformed Real FDI"]][["6a"]] <- model.6a
hsd_testing[["IHS-Transformed Real FDI"]][["6a"]] <- bptest(model.6a)

model.7a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r) + lag(`Inflation.GDP.deflator.annual.pct`), data = data)
models[["IHS-Transformed Real FDI"]][["7a"]] <- model.7a
hsd_testing[["IHS-Transformed Real FDI"]][["7a"]] <- bptest(model.7a)

model.8a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r) + lag(`Inflation.GDP.deflator.annual.pct`) + lag(GDP.growth.annual.pct), data = data)
models[["IHS-Transformed Real FDI"]][["8a"]] <- model.8a
hsd_testing[["IHS-Transformed Real FDI"]][["8a"]] <- bptest(model.8a)
```

```{r}
models_unlisted <- unlist(models, recursive=FALSE)
names(models_unlisted) <- c('1a', '2a', '3a', '4a', '5a', '6a', '7a', '8a')

cm <- c( '(Intercept)' = 'Constant', 'passengers.in.change.pct' = 'Change in passengers (%)',
         'lag(log_battle)' = 'Battle fatalities (log)', 'lag(log_GDP_r)' = 'Real GDP (int$, log, lag)',
         'lag(log_GDP_r_pc)' = 'Real GDP per capita (int$, log, lag)',
         'lag(log1p(total_GDP_n_ID))' = 'Nominal GDP (int$, log, lag)',
         'lag(log1p(`GDP.current.US$`))' = 'Nominal GDP (USD, log, lag)',
         'lag(Inflation.GDP.deflator.annual.pct)' = 'Inflation GDP Deflator',
         'lag(GDP.growth.annual.pct)' = 'GDP Growth (%, lag)')

# Generate the model summary
output <- msummary(
  models_unlisted, output = 'kableExtra', coef_map = cm, stars = TRUE, 
  gof_omit = 'IC|Log|Adj', title = "IHS-Transfomed Real FDI", 
  statistic = 'p.value', conf_level = 0.95, vcov = 'robust'
)

# Save the table as an HTML file
save_kable(output, file = "Multiple OLS Analysis/IHS-summary_ALL.html")
```
Source: https://stackoverflow.com/questions/62610249/modelsummary-kableextra-regression-table-with-models-of-the-same-name 




##Europe only

Just European countries:
```{r}
data_Eur <- data %>%
  filter(Country.Name %in% c("Armenia", "Austria", "Belarus", "Belgium", "Bosnia and Herzegovina", "Bulgaria", "Croatia", "Cyprus", "Czechia", "Denmark", "Estonia", "Finland", "France", "Georgia", "Germany", "Greece", "Iceland", "Ireland", "Italy", "Latvia", "Lithuania", "Luxembourg", "Malta", "Moldova", "Netherlands", "North Macedonia", "Norway", "Poland", "Portugal", "Romania", "Russia", "Serbia", "Slovenia", "Spain", "Sweden", "Switzerland",  "Turkiye", "Ukraine", "United Kingdom"))
unique(data_Eur$Country.Name)
```

IHS-Transformed Real FDI
```{r}
models <- NA
models <- list()
hsd_testing <- NA
hsd_testing <- list()

model.1a <- lm(ihs_FDI_r ~ passengers.in.change.pct, data = data_Eur)
models[["IHS-Transformed Real FDI"]][["1a"]] <- model.1a
hsd_testing[["IHS-Transformed Real FDI"]][["1a"]] <- bptest(model.1a)

model.2a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle), data = data_Eur)
models[["IHS-Transformed Real FDI"]][["2a"]] <- model.2a
hsd_testing[["IHS-Transformed Real FDI"]][["2a"]] <- bptest(model.2a)

model.3a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r), data = data_Eur)
models[["IHS-Transformed Real FDI"]][["3a"]] <- model.3a
hsd_testing[["IHS-Transformed Real FDI"]][["3a"]] <- bptest(model.3a)

model.4a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r_pc), data = data_Eur)
models[["IHS-Transformed Real FDI"]][["4a"]] <- model.4a
hsd_testing[["IHS-Transformed Real FDI"]][["4a"]] <- bptest(model.4a)

model.5a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log1p(total_GDP_n_ID)), data = data_Eur)
models[["IHS-Transformed Real FDI"]][["5a"]] <- model.5a
hsd_testing[["IHS-Transformed Real FDI"]][["5a"]] <- bptest(model.5a)

model.6a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log1p(`GDP.current.US$`)), data = data_Eur)
models[["IHS-Transformed Real FDI"]][["6a"]] <- model.6a
hsd_testing[["IHS-Transformed Real FDI"]][["6a"]] <- bptest(model.6a)

model.7a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r) + lag(`Inflation.GDP.deflator.annual.pct`), data = data_Eur)
models[["IHS-Transformed Real FDI"]][["7a"]] <- model.7a
hsd_testing[["IHS-Transformed Real FDI"]][["7a"]] <- bptest(model.7a)

model.8a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r) + lag(`Inflation.GDP.deflator.annual.pct`) + lag(GDP.growth.annual.pct), data = data_Eur)
models[["IHS-Transformed Real FDI"]][["8a"]] <- model.8a
hsd_testing[["IHS-Transformed Real FDI"]][["8a"]] <- bptest(model.8a)
```

```{r}
models_unlisted <- unlist(models, recursive=FALSE)
names(models_unlisted) <- c('1a', '2a', '3a', '4a', '5a', '6a', '7a', '8a')

cm <- c( '(Intercept)' = 'Constant', 'passengers.in.change.pct' = 'Change in passengers (%)',
         'lag(log_battle)' = 'Battle fatalities (log)', 'lag(log_GDP_r)' = 'Real GDP (int$, log, lag)',
         'lag(log_GDP_r_pc)' = 'Real GDP per capita (int$, log, lag)',
         'lag(log1p(total_GDP_n_ID))' = 'Nominal GDP (int$, log, lag)',
         'lag(log1p(`GDP.current.US$`))' = 'Nominal GDP (USD, log, lag)',
         'lag(Inflation.GDP.deflator.annual.pct)' = 'Inflation GDP Deflator',
         'lag(GDP.growth.annual.pct)' = 'GDP Growth (%, lag)')

# Generate the model summary
output <- msummary(
  models_unlisted, output = 'kableExtra', coef_map = cm, stars = TRUE, 
  gof_omit = 'IC|Log|Adj', title = "IHS-Transfomed Real FDI", 
  statistic = 'p.value', conf_level = 0.95, vcov = 'robust'
)

# Save the table as an HTML file
save_kable(output, file = "Multiple OLS Analysis/IHS-summary_EUR.html")
```




##LM countries 



***Just Middle Income Countries***

Just Middle Income countries:
```{r}
data_MI <- data %>%
  filter(Country.Name %in% c("Algeria", "Angola", "Bangladesh", "Benin", "Bhutan", "Bolivia", "Cabo Verde", "Cambodia", "Cameroon", "Comoros", "Congo, Rep.", "Côte d'Ivoire", "Djibouti", "Egypt", "Eswatini", "Ghana", "Guinea", "Haiti", "Honduras", "India", "Iran", "Jordan", "Kenya", "Kiribati", "Kyrgyz Republic", "Lao PDR", "Lebanon", "Lesotho", "Mauritania", "Micronesia", "Mongolia", "Morocco", "Myanmar", "Nepal", "Nicaragua", "Nigeria", "Pakistan", "Papua New Guinea", "Philippines", "Samoa", "São Tomé and Príncipe", "Senegal", "Solomon Islands", "Sri Lanka", "Tajikistan", "Tanzania", "Timor-Leste", "Tunisia", "Ukraine", "Uzbekistan", "Vanuatu", "Vietnam", "Zambia", "Zimbabwe", "Albania", "Argentina", "Armenia", "Azerbaijan", "Belarus", "Belize", "Bosnia and Herzegovina", "Botswana", "Brazil", "Bulgaria", "China", "Colombia", "Costa Rica", "Cuba", "Dominica", "Dominican Republic", "Ecuador", "El Salvador", "Equatorial Guinea", "Fiji", "Gabon", "Georgia", "Grenada", "Guatemala", "Indonesia", "Iraq", "Jamaica", "Kazakhstan", "Kosovo", "Libya", "Malaysia", "Maldives", "Marshall Islands", "Mauritius", "Mexico", "Moldova", "Montenegro", "Namibia", "North Macedonia", "Palau", "Paraguay", "Peru", "Russian Federation", "Serbia", "South Africa", "St. Lucia", "St. Vincent and the Grenadines", "Suriname", "Thailand", "Tonga", "Türkiye", "Turkmenistan", "Tuvalu"))
unique(data_MI$Country.Name)
```


IHS-Transformed Real FDI
```{r}
models <- NA
models <- list()
hsd_testing <- NA
hsd_testing <- list()

model.1a <- lm(ihs_FDI_r ~ passengers.in.change.pct, data = data_MI)
models[["IHS-Transformed Real FDI"]][["1a"]] <- model.1a
hsd_testing[["IHS-Transformed Real FDI"]][["1a"]] <- bptest(model.1a)

model.2a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle), data = data_MI)
models[["IHS-Transformed Real FDI"]][["2a"]] <- model.2a
hsd_testing[["IHS-Transformed Real FDI"]][["2a"]] <- bptest(model.2a)

model.3a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r), data = data_MI)
models[["IHS-Transformed Real FDI"]][["3a"]] <- model.3a
hsd_testing[["IHS-Transformed Real FDI"]][["3a"]] <- bptest(model.3a)

model.4a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r_pc), data = data_MI)
models[["IHS-Transformed Real FDI"]][["4a"]] <- model.4a
hsd_testing[["IHS-Transformed Real FDI"]][["4a"]] <- bptest(model.4a)

model.5a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log1p(total_GDP_n_ID)), data = data_MI)
models[["IHS-Transformed Real FDI"]][["5a"]] <- model.5a
hsd_testing[["IHS-Transformed Real FDI"]][["5a"]] <- bptest(model.5a)

model.6a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log1p(`GDP.current.US$`)), data = data_MI)
models[["IHS-Transformed Real FDI"]][["6a"]] <- model.6a
hsd_testing[["IHS-Transformed Real FDI"]][["6a"]] <- bptest(model.6a)

model.7a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r) + lag(`Inflation.GDP.deflator.annual.pct`), data = data_MI)
models[["IHS-Transformed Real FDI"]][["7a"]] <- model.7a
hsd_testing[["IHS-Transformed Real FDI"]][["7a"]] <- bptest(model.7a)

model.8a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r) + lag(`Inflation.GDP.deflator.annual.pct`) + lag(GDP.growth.annual.pct), data = data_MI)
models[["IHS-Transformed Real FDI"]][["8a"]] <- model.8a
hsd_testing[["IHS-Transformed Real FDI"]][["8a"]] <- bptest(model.8a)
```

```{r}
models_unlisted <- unlist(models, recursive=FALSE)
names(models_unlisted) <- c('1a', '2a', '3a', '4a', '5a', '6a', '7a', '8a')

cm <- c( '(Intercept)' = 'Constant', 'passengers.in.change.pct' = 'Change in passengers (%)',
         'lag(log_battle)' = 'Battle fatalities (log)', 'lag(log_GDP_r)' = 'Real GDP (int$, log, lag)',
         'lag(log_GDP_r_pc)' = 'Real GDP per capita (int$, log, lag)',
         'lag(log1p(total_GDP_n_ID))' = 'Nominal GDP (int$, log, lag)',
         'lag(log1p(`GDP.current.US$`))' = 'Nominal GDP (USD, log, lag)',
         'lag(Inflation.GDP.deflator.annual.pct)' = 'Inflation GDP Deflator',
         'lag(GDP.growth.annual.pct)' = 'GDP Growth (%, lag)')

# Generate the model summary
output <- msummary(
  models_unlisted, output = 'kableExtra', coef_map = cm, stars = TRUE, 
  gof_omit = 'IC|Log|Adj', title = "IHS-Transfomed Real FDI", 
  statistic = 'p.value', conf_level = 0.95, vcov = 'robust'
)

# Save the table as an HTML file
save_kable(output, file = "Multiple OLS Analysis/IHS-summary_MIC.html")
```





***Just Lower Middle Income Countries***

Just Lower Middle Income countries:
```{r}
data_LMI <- data %>%
  filter(Country.Name %in% c("Algeria", "Angola", "Bangladesh", "Benin", "Bhutan", "Bolivia", "Cabo Verde", "Cambodia", "Cameroon", "Comoros", "Congo, Rep.", "Côte d'Ivoire", "Djibouti", "Egypt", "Eswatini", "Ghana", "Guinea", "Haiti", "Honduras", "India", "Iran", "Jordan", "Kenya", "Kiribati", "Kyrgyz Republic", "Lao PDR", "Lebanon", "Lesotho", "Mauritania", "Micronesia", "Mongolia", "Morocco", "Myanmar", "Nepal", "Nicaragua", "Nigeria", "Pakistan", "Papua New Guinea", "Philippines", "Samoa", "São Tomé and Príncipe", "Senegal", "Solomon Islands", "Sri Lanka", "Tajikistan", "Tanzania", "Timor-Leste", "Tunisia", "Ukraine", "Uzbekistan", "Vanuatu", "Vietnam", "Zambia", "Zimbabwe"))
unique(data_LMI$Country.Name)
```

IHS-Transformed Real FDI
```{r}
models <- NA
models <- list()
hsd_testing <- NA
hsd_testing <- list()

model.1a <- lm(ihs_FDI_r ~ passengers.in.change.pct, data = data_LMI)
models[["IHS-Transformed Real FDI"]][["1a"]] <- model.1a
hsd_testing[["IHS-Transformed Real FDI"]][["1a"]] <- bptest(model.1a)

model.2a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle), data = data_LMI)
models[["IHS-Transformed Real FDI"]][["2a"]] <- model.2a
hsd_testing[["IHS-Transformed Real FDI"]][["2a"]] <- bptest(model.2a)

model.3a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r), data = data_LMI)
models[["IHS-Transformed Real FDI"]][["3a"]] <- model.3a
hsd_testing[["IHS-Transformed Real FDI"]][["3a"]] <- bptest(model.3a)

model.4a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r_pc), data = data_LMI)
models[["IHS-Transformed Real FDI"]][["4a"]] <- model.4a
hsd_testing[["IHS-Transformed Real FDI"]][["4a"]] <- bptest(model.4a)

model.5a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log1p(total_GDP_n_ID)), data = data_LMI)
models[["IHS-Transformed Real FDI"]][["5a"]] <- model.5a
hsd_testing[["IHS-Transformed Real FDI"]][["5a"]] <- bptest(model.5a)

model.6a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log1p(`GDP.current.US$`)), data = data_LMI)
models[["IHS-Transformed Real FDI"]][["6a"]] <- model.6a
hsd_testing[["IHS-Transformed Real FDI"]][["6a"]] <- bptest(model.6a)

model.7a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r) + lag(`Inflation.GDP.deflator.annual.pct`), data = data_LMI)
models[["IHS-Transformed Real FDI"]][["7a"]] <- model.7a
hsd_testing[["IHS-Transformed Real FDI"]][["7a"]] <- bptest(model.7a)

model.8a <- lm(ihs_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r) + lag(`Inflation.GDP.deflator.annual.pct`) + lag(GDP.growth.annual.pct), data = data_LMI)
models[["IHS-Transformed Real FDI"]][["8a"]] <- model.8a
hsd_testing[["IHS-Transformed Real FDI"]][["8a"]] <- bptest(model.8a)
```

```{r}
models_unlisted <- unlist(models, recursive=FALSE)
names(models_unlisted) <- c('1a', '2a', '3a', '4a', '5a', '6a', '7a', '8a')

cm <- c( '(Intercept)' = 'Constant', 'passengers.in.change.pct' = 'Change in passengers (%)',
         'lag(log_battle)' = 'Battle fatalities (log)', 'lag(log_GDP_r)' = 'Real GDP (int$, log, lag)',
         'lag(log_GDP_r_pc)' = 'Real GDP per capita (int$, log, lag)',
         'lag(log1p(total_GDP_n_ID))' = 'Nominal GDP (int$, log, lag)',
         'lag(log1p(`GDP.current.US$`))' = 'Nominal GDP (USD, log, lag)',
         'lag(Inflation.GDP.deflator.annual.pct)' = 'Inflation GDP Deflator',
         'lag(GDP.growth.annual.pct)' = 'GDP Growth (%, lag)')

# Generate the model summary
output <- msummary(
  models_unlisted, output = 'kableExtra', coef_map = cm, stars = TRUE, 
  gof_omit = 'IC|Log|Adj', title = "IHS-Transfomed Real FDI", 
  statistic = 'p.value', conf_level = 0.95, vcov = 'robust'
)

# Save the table as an HTML file
save_kable(output, file = "Multiple OLS Analysis/IHS-summary_LMIC.html")
```






##Log transformations



Log-Transformed Real FDI
```{r}
models <- NA
models <- list()
hsd_testing <- NA
hsd_testing <- list()

model.1a <- lm(log_FDI_r ~ passengers.in.change.pct, data = data)
models[["Log-Transformed Real FDI"]][["1a"]] <- model.1a
hsd_testing[["Log-Transformed Real FDI"]][["1a"]] <- bptest(model.1a)

model.2a <- lm(log_FDI_r ~ passengers.in.change.pct + lag(log_battle), data = data)
models[["Log-Transformed Real FDI"]][["2a"]] <- model.2a
hsd_testing[["Log-Transformed Real FDI"]][["2a"]] <- bptest(model.2a)

model.3a <- lm(log_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r), data = data)
models[["Log-Transformed Real FDI"]][["3a"]] <- model.3a
hsd_testing[["Log-Transformed Real FDI"]][["3a"]] <- bptest(model.3a)

model.4a <- lm(log_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r_pc), data = data)
models[["Log-Transformed Real FDI"]][["4a"]] <- model.4a
hsd_testing[["Log-Transformed Real FDI"]][["4a"]] <- bptest(model.4a)

model.5a <- lm(log_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log1p(total_GDP_n_ID)), data = data)
models[["Log-Transformed Real FDI"]][["5a"]] <- model.5a
hsd_testing[["Log-Transformed Real FDI"]][["5a"]] <- bptest(model.5a)

model.6a <- lm(log_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log1p(`GDP.current.US$`)), data = data)
models[["Log-Transformed Real FDI"]][["6a"]] <- model.6a
hsd_testing[["Log-Transformed Real FDI"]][["6a"]] <- bptest(model.6a)

model.7a <- lm(log_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r) + lag(`Inflation.GDP.deflator.annual.pct`), data = data)
models[["Log-Transformed Real FDI"]][["7a"]] <- model.7a
hsd_testing[["Log-Transformed Real FDI"]][["7a"]] <- bptest(model.7a)

model.8a <- lm(log_FDI_r ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r) + lag(`Inflation.GDP.deflator.annual.pct`) + lag(GDP.growth.annual.pct), data = data)
models[["Log-Transformed Real FDI"]][["8a"]] <- model.8a
hsd_testing[["Log-Transformed Real FDI"]][["8a"]] <- bptest(model.8a)
```

```{r}
models_unlisted <- unlist(models, recursive=FALSE)
names(models_unlisted) <- c('1a', '2a', '3a', '4a', '5a', '6a', '7a', '8a')

cm <- c( '(Intercept)' = 'Constant', 'passengers.in.change.pct' = 'Change in passengers (%)',
         'lag(log_battle)' = 'Battle fatalities (log)', 'lag(log_GDP_r)' = 'Real GDP (int$, log, lag)',
         'lag(log_GDP_r_pc)' = 'Real GDP per capita (int$, log, lag)',
         'lag(log1p(total_GDP_n_ID))' = 'Nominal GDP (int$, log, lag)',
         'lag(log1p(`GDP.current.US$`))' = 'Nominal GDP (USD, log, lag)',
         'lag(Inflation.GDP.deflator.annual.pct)' = 'Inflation GDP Deflator',
         'lag(GDP.growth.annual.pct)' = 'GDP Growth (%, lag)')

msummary(models_unlisted, output = 'kableExtra', coef_map = cm, stars = TRUE, gof_omit = 'IC|Log|Adj', title = "Log-Transfomed Real FDI", statistic = 'p.value', conf_level = 0.95, vcov = 'robust')
```

Log-Transformed Nominal FDI
```{r}
models <- NA
models <- list()
hsd_testing <- NA
hsd_testing <- list()

model.1a <- lm(log_FDI_n ~ passengers.in.change.pct, data = data)
models[["Log-Transformed Nominal FDI"]][["1a"]] <- model.1a
hsd_testing[["Log-Transformed Nominal FDI"]][["1a"]] <- bptest(model.1a)

model.2a <- lm(log_FDI_n ~ passengers.in.change.pct + lag(log_battle), data = data)
models[["Log-Transformed Nominal FDI"]][["2a"]] <- model.2a
hsd_testing[["Log-Transformed Nominal FDI"]][["2a"]] <- bptest(model.2a)

model.3a <- lm(log_FDI_n ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r), data = data)
models[["Log-Transformed Nominal FDI"]][["3a"]] <- model.3a
hsd_testing[["Log-Transformed Nominal FDI"]][["3a"]] <- bptest(model.3a)

model.4a <- lm(log_FDI_n ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r_pc), data = data)
models[["Log-Transformed Nominal FDI"]][["4a"]] <- model.4a
hsd_testing[["Log-Transformed Nominal FDI"]][["4a"]] <- bptest(model.4a)

model.5a <- lm(log_FDI_n ~ passengers.in.change.pct + lag(log_battle) + lag(log1p(total_GDP_n_ID)), data = data)
models[["Log-Transformed Nominal FDI"]][["5a"]] <- model.5a
hsd_testing[["Log-Transformed Nominal FDI"]][["5a"]] <- bptest(model.5a)

model.6a <- lm(log_FDI_n ~ passengers.in.change.pct + lag(log_battle) + lag(log1p(`GDP.current.US$`)), data = data)
models[["Log-Transformed Nominal FDI"]][["6a"]] <- model.6a
hsd_testing[["Log-Transformed Nominal FDI"]][["6a"]] <- bptest(model.6a)

model.7a <- lm(log_FDI_n ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r) + lag(`Inflation.GDP.deflator.annual.pct`), data = data)
models[["Log-Transformed Nominal FDI"]][["7a"]] <- model.7a
hsd_testing[["Log-Transformed Nominal FDI"]][["7a"]] <- bptest(model.7a)

model.8a <- lm(log_FDI_n ~ passengers.in.change.pct + lag(log_battle) + lag(log_GDP_r) + lag(`Inflation.GDP.deflator.annual.pct`) + lag(GDP.growth.annual.pct), data = data)
models[["Log-Transformed Nominal FDI"]][["8a"]] <- model.8a
hsd_testing[["Log-Transformed Nominal FDI"]][["8a"]] <- bptest(model.8a)
```

```{r}
models_unlisted <- unlist(models, recursive=FALSE)
names(models_unlisted) <- c('1a', '2a', '3a', '4a', '5a', '6a', '7a', '8a')

cm <- c( '(Intercept)' = 'Constant', 'passengers.in.change.pct' = 'Change in passengers (%)',
         'lag(log_battle)' = 'Battle fatalities (log)', 'lag(log_GDP_r)' = 'Real GDP (int$, log, lag)',
         'lag(log_GDP_r_pc)' = 'Real GDP per capita (int$, log, lag)',
         'lag(log1p(total_GDP_n_ID))' = 'Nominal GDP (int$, log, lag)',
         'lag(log1p(`GDP.current.US$`))' = 'Nominal GDP (USD, log, lag)',
         'lag(Inflation.GDP.deflator.annual.pct)' = 'Inflation GDP Deflator',
         'lag(GDP.growth.annual.pct)' = 'GDP Growth (%, lag)')

msummary(models_unlisted, output = 'kableExtra', coef_map = cm, stars = TRUE, gof_omit = 'IC|Log|Adj', title = "Log-Transfomed Nominal FDI", statistic = 'p.value', conf_level = 0.95, vcov = 'robust')
```
